Peek at value of expression

This function is a proxy for calling peeking function.

ic_peek(
  value,
  peeking_function = getOption("icecream.peeking.function"),
  max_lines = getOption("icecream.max.lines")
)

Arguments

value
The result of evaluating an expression inside the ic() function.

peeking_function
The function used to peek at the value. Default value is set by the "icecream.peeking.function" option.

max_lines
Maximum number of lines printed. Default value is set by the "icecream.max.lines" option.

Value

A string to be printed.

Details

Default value of icecream.peeking.function is ic_autopeek. Suggested possible alternatives are:

utils::str

print

head

summary

tibble::glimpse

See also