Icecream: Never use print() to debug again

Icecream provides a more user-friendly print debugging experience. Use ic() wherever you would use print() and see the expression and its value easily.

Options

The following options can be used to control behaviour:

icecream.enabled: Boolean. If FALSE, calls to ic(foo) simply evaluate and return foo. No output is printed.

icecream.prefix: This is printed at the beginning of every line. Defaults to "ic|".

icecream.include.context: Boolean. If TRUE, when calling ic(foo) the source file:line or environment will be printed along with the expression and value. This can be useful for more complicated debugging but produces a lot of output so is disabled by default.

icecream.output.function: Not implemented yet. See the configuration section of the original project docs for details of what it will do.

icecream.arg.to.string.function: Not implemented yet. See the configuration section of the original project docs for details of what it will do.